layers: Add Copy Blit Resolve CoreCheck file
